Create Break the Door
Problem

Don has n hammers. The hammer i has attack power Pi. He needs to destroy a wooden door, the strength of which is d. When a hammer with power Pi strikes the door, the strength of the door is decreased by Pi (if d<pi, then d becomes 0) and the attack power of the hammer becomes floor(Pi/2).
Don has k trials, and he has to break the doors within those k trials.
If it is not possible, print -1.
If possible, print the least number of hits required.